pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[project]
name = "iq-sdk"
version = "0.1.0"
authors = [
  { name="Tianshu Huang", email="tianshu2@andrew.cmu.edu" },
]
description = "TODO"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
license-files = ["LICEN[CS]E*"]
classifiers = [
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Typing :: Typed",
    "Operating System :: POSIX :: Linux",
]

dependencies = [
    "abstract-dataloader>=0.1",
    "jaxtyping>=0.3.9",
    "numpy>=1.14",
    "pyyaml>=6.0",
    "wadler-lindig>=0.1.7",
]

[project.urls]
Homepage = "https://deep-spectrum.github.io/iq-sdk"
Documentation = "https://deep-spectrum.github.io/iq-sdk"
Issues = "https://github.com/deep-spectrum/iq-sdk/issues"

[project.optional-dependencies]
dev = [
    "pre-commit >= 4.0.0",
    "pytest >= 8.0.0",
    "coverage",
    "pytest-cov",
    "pyright >= 1.0.0",
    "ruff >= 0.11.0"
]
docs = [
    "mkdocs == 1.6.*",
    "mkdocs-material == 9.6.*",
    "mkdocstrings-python-xref == 1.16.3",
    "mkdocstrings-python == 1.16.6",
    "griffe == 1.8.0",
    "ruff >= 0.11.0",
    "click==8.2.1"
]
dsp = [
    "jax[cuda13]>=0.9.0.1",
]
debug = [
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "pandas>=3.0.1",
    "scipy>=1.17.0",
    "ipykernel>=7.2.0",
]

[tool.ruff]
line-length = 80
indent-width = 4

[tool.ruff.lint]
select = [
    "E", "F", "W", "N", "I", "D", "NPY"
]
ignore = [
    "D102",  # Inheriting docstrings is allowed
    "D107",  # Initializer goes in the class docstring
    "D401",  # "Imperative mood" NLP doesn't always make sense
    "D413",  # No blank lines after section is preferred by mkdocstrings
    "N806",  # We do math, so capitalization should be allowed
    "E402",  # Module level imports sometimes need to occur after configuration
    "F722",  # Jaxtyping annotations confuse ruff
    "F821",

    # Must pick one...
    "D213",  # Pick 'multi-line-summary-first-line'.
    "D203",  # Pick `no-blank-line-before-class`.
]

[tool.coverage.report]
exclude_also = [
    '^\s*\.\.\.\s*$',
    'def __repr__',
    'if __name__ == .__main__.:',
    'class .*\bProtocol\):',
    'raise NotImplementedError',
]
